jsl.js lazy loading.
Improved code so jsl.js (and all the related features) are lazy-loaded.
Added loader.html to test the lazy loading.

Code needs now to be adapted a little bit as the exceptions due to jOVI
not being initialized are no longer being thrown as methods calls are
just queued when jOVI hasn't been initialized in the target container
ready to be executed upon initialization. This is not necessarily a bad
behavior.

Note that because of the way the JSLA loader works it has to be
triggered on the onload event as opposed of the ondomready event.
Getting started with the wrong event listener just empties the DOM. God
knows why...